New pipeline post dj created
When I was in the process of checking why we were getting noise through the dj pipeline. I created a new secondary post-dj pipeline to change parameters quickly without the need to reingest.

My plan is to ultimately change the re-ingestion pipeline to match the parameters I found work best. but for now im loading the single trial data and analyzing it through new functions.
Happy to walk you through it. The functions are all readable and modular and commented. I know it's a hassle to have a new set of functions but it has sped up my progress and trouble shooting significantly since I'm much more familiar  with the code.

Right now the inclusion parameters that i'm using  for TrigDffInclusionParams is 9:
(9, 1.96, 0.6, 1, 'peak', 'std', 25, True)
From this the one that 25 um distance is being used for is_included True is being used so only somas are being analyzed.
The rest of the parameters are being implemented after I use get_single_trial_data to fetch trial data.

 The parameters for responding cells are:-z-scored dFF
-Medfilt (7 points)-single trial 2.94 z-score amplitude cutoff-min_width_val=2 points-response_proportion >= 0.6-peak_time_std <= 0.75